requirements.txt
# Backend API (api.py)
flask>=3.0
flask-cors>=4.0
google-generativeai>=0.3.0
